feat(choose-plan-page): Add imagesFormat configuration to PaymentPage interactor
-  Introduce imagesFormat configuration to customize supported image formats in PaymentPage interactor
