+++ src/modules/choose-plan-page/interactor.config.ts
@@ -1,0 +1,14 @@
+import { InternalFileType } from '../../shared/types';
+
+export const interactorImagesFormat = [
+  InternalFileType.HEIC,
+  InternalFileType.SVG,
+  InternalFileType.PNG,
+  InternalFileType.BMP,
+  InternalFileType.EPS,
+  InternalFileType.GIF,
+  InternalFileType.TIFF,
+  InternalFileType.WEBP,
+  InternalFileType.JPG,
+  InternalFileType.JPEG,
+] as const satisfies InternalFileType[];
