+++ src/shared/types.ts
@@ -1,0 +1,22 @@
+export enum InternalFileType {
+  DOC = "DOC",
+  DOCX = "DOCX",
+  JPEG = "JPEG",
+  JPG = "JPG",
+  HEIC = "HEIC",
+  HEIF = "HEIF",
+  PDF = "PDF",
+  PNG = "PNG",
+  PPT = "PPT",
+  PPTX = "PPTX",
+  XLS = "XLS",
+  XLSX = "XLSX",
+  ZIP = "ZIP",
+  BMP = "BMP",
+  EPS = "EPS",
+  GIF = "GIF",
+  SVG = "SVG",
+  TIFF = "TIFF",
+  WEBP = "WEBP",
+  EPUB = "EPUB",
+}
